refactor: Move tests and fix imports
